追加機能実装（WebSocketを使用した機能）

+++ src/main/java/com/example/forum_hitsumabushi/controller/form/LikeForm.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class LikeForm {
     private String targetType;
-    private Long targetId;
+    private Integer targetId;
 }
